adminpanel-ms/requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nats-py==2.8.0
python-dotenv==1.0.1
yarl==1.9.4
motor==3.5.1
pymongo==4.8.0
dnspython==2.6.1
